# pkg/client/openapi.yaml
components:
  schemas:
...
    ContactState:
      properties:
        cadence_due_at:
          format: date-time
          type: string
        cadence_status:
          type: string
        computed_at:
          format: date-time
          type: string
        first_contact_at:
          format: date-time
          type: string
        first_contact_ref:
          type: string
        inferred_channel:
          type: string
        interaction_count:
          format: int64
          type: integer
        last_contact_at:
          format: date-time
          type: string
        last_contact_channel:
          type: string
        last_contact_owner:
          type: string
        last_contact_ref:
          type: string
        last_contact_source_id:
          format: int64
          type: integer
        last_inbound_at:
          format: date-time
          type: string
        last_inbound_ref:
          type: string
        last_outbound_at:
          format: date-time
          type: string
        last_outbound_ref:
          type: string
        person_id:
          format: int64
          type: integer
        stale:
          type: boolean
      required:
        - person_id
        - interaction_count
        - cadence_status
        - stale
        - computed_at
      type: object
...
    CreateDailyNoteEntryRequest:
      additionalProperties: false
      properties:
        author:
          type: string
        body:
          type: string
        person_ids:
          items:
            format: int64
            minimum: 1
            type: integer
          nullable: true
          type: array
      required:
        - body
      type: object
...
    DailyNoteEntriesResponse:
      properties:
        entries:
          items:
            $ref: "#/components/schemas/DailyNoteEntry"
          type: array
      required:
        - entries
      type: object
    DailyNoteEntry:
      properties:
        author:
          type: string
        body:
          type: string
        created_at:
          format: date-time
          type: string
        id:
          format: int64
          type: integer
        local_date:
          type: string
        ordinal:
          format: int64
          type: integer
        person_ids:
          items:
            format: int64
            type: integer
          nullable: true
          type: array
        source:
          type: string
        source_ref:
          type: string
        updated_at:
          format: date-time
          type: string
      required:
        - id
        - local_date
        - ordinal
        - body
        - author
        - source
        - person_ids
        - created_at
        - updated_at
      type: object
    DayPage:
      properties:
        entries:
          items:
            $ref: "#/components/schemas/DailyNoteEntry"
          type: array
        entry_total_count:
          format: int64
          type: integer
        local_date:
          type: string
        person_total_count:
          format: int64
          type: integer
        persons:
          items:
            $ref: "#/components/schemas/DayPerson"
          type: array
      required:
        - local_date
        - persons
        - entries
        - person_total_count
        - entry_total_count
      type: object
...
    ErrorResponse:
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: X-Api-Key
      type: apiKey
info:
  title: msgvault API
  version: 1.33.0
openapi: 3.0.3
paths:
...
  /api/v1/days/entries/{id}:
    delete:
      operationId: deleteDayEntry
      parameters:
        - description: Positive durable identifier
          in: path
          name: id
          required: true
          schema:
            format: int64
            minimum: 1
            type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          description: Error
      security:
        - apiKey: []
      summary: Delete an authored daily entry
      tags:
        - API
  /api/v1/days/{date}:
    get:
      operationId: getActivityDay
      parameters:
        - description: Exact local calendar date
          in: path
          name: date
          required: true
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
        - description: Maximum primary rows to return
          in: query
          name: limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based primary row offset
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
        - description: Maximum authored entries to return independently
          in: query
          name: entry_limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based authored-entry offset
          in: query
          name: entry_offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
        - description: Maximum activity references previewed for each person
          in: query
          name: activity_limit_per_person
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DayPage"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get people, activity, and notes for a day
      tags:
        - API
  /api/v1/days/{date}/entries:
    get:
      operationId: listDayEntries
      parameters:
        - description: Exact local calendar date
          in: path
          name: date
          required: true
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
        - description: Maximum primary rows to return
          in: query
          name: limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based primary row offset
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyNoteEntriesResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: List authored entries for a day
      tags:
        - API
    post:
      operationId: createDayEntry
      parameters:
        - description: Exact local calendar date
          in: path
          name: date
          required: true
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyNoteEntryRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyNoteEntry"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "413":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Create an authored entry for a day
      tags:
        - API
...
  /api/v1/persons/{id}/contact-state:
    get:
      operationId: getPersonContactState
      parameters:
        - description: Positive durable identifier
          in: path
          name: id
          required: true
          schema:
            format: int64
            minimum: 1
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactState"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get computed contact state for a person
      tags:
        - API
  /api/v1/persons/{id}/days:
    get:
      operationId: listPersonActivityDays
      parameters:
        - description: Positive durable identifier
          in: path
          name: id
          required: true
          schema:
            format: int64
            minimum: 1
            type: integer
        - description: Inclusive first local calendar date
          in: query
          name: from
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
        - description: Inclusive last local calendar date
          in: query
          name: to
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
        - description: Maximum primary rows to return
          in: query
          name: limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based primary row offset
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDaysPage"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: List calendar days intersecting a person
      tags:
        - API
  /api/v1/persons/{id}/days/{date}:
    get:
      operationId: getPersonActivityDay
      parameters:
        - description: Positive durable identifier
          in: path
          name: id
          required: true
          schema:
            format: int64
            minimum: 1
            type: integer
        - description: Exact local calendar date
          in: path
          name: date
          required: true
          schema:
            pattern: ^[0-9]{4}-[0-9]{2}-[0-9]{2}$
            type: string
        - description: Maximum primary rows to return
          in: query
          name: limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based primary row offset
          in: query
          name: offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
        - description: Maximum authored entries to return independently
          in: query
          name: entry_limit
          schema:
            default: 100
            format: int64
            maximum: 500
            minimum: 1
            type: integer
        - description: Zero-based authored-entry offset
          in: query
          name: entry_offset
          schema:
            default: 0
            format: int64
            minimum: 0
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonDayPage"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Error
      security:
        - apiKey: []
      summary: Get one person's activity and notes for a day
      tags:
        - API
...